requirements.txt
Django==1.10.5
PyJWT==1.4.2
django-cors-headers==2.0.2
django-filter==1.0.1
djangorestframework==3.5.3
djangorestframework-jwt==1.9.0
djoser==0.5.4
dry-rest-permissions==0.1.8
six==1.10.0
